add user and link
