+++ .gitignore
@@ -1,1 +1,1 @@
-./api/fostercollab-api
+./API/fostercollab-api
